Added mysql:find-replace command and task (task #4292)
Added mysql:find-replace command and task.
Note that on PHP 7.1 there is a problem with
interconnectit/search-replace-db that gives
notice on non well formed numeric value and
doesn't do a report, while actual replacement
takes place. This is due to new return value
in time related functions of new PHP