Disable strict composer validation
Composer validation is useful for catching syntax errors and
the like.  However within WordPress ecosystem and WPackagist,
it's not feasible to use non-specific versions, as WordPress
community does not always follow semantic versioning.

Strict composer validation fails on the warnings as well, so
instead of completely removing the validation, we set it to
be non-strict.